+++ tests/TestCase/Http/ErrorHandlerTest.php
@@ -200,7 +200,7 @@ class ErrorHandlerTest extends \PHPUnit\Framework\TestCase
     public function testExceptionHandlerAjax()
     {
         $request = new Request('/api/users/fetch');
-        $request->type('json');
+        $request->header('Accept', 'application/json');
         Router::request($request);
 
         $errorHandler = new MockErrorHandler();
@@ -218,7 +218,7 @@ class ErrorHandlerTest extends \PHPUnit\Framework\TestCase
     public function testExceptionHandlerAjaxDebugDisabled()
     {
         $request = new Request('/api/users/fetch');
-        $request->type('json');
+        $request->header('Accept', 'application/json');
         Router::request($request);
 
         $errorHandler = new MockErrorHandler();
@@ -236,7 +236,7 @@ class ErrorHandlerTest extends \PHPUnit\Framework\TestCase
     public function testExceptionHandlerAjaxDebugDisabledNonHttp()
     {
         $request = new Request('/api/users/fetch');
-        $request->type('json');
+        $request->header('Accept', 'application/json');
         Router::request($request);
 
         $errorHandler = new MockErrorHandler();
@@ -254,7 +254,7 @@ class ErrorHandlerTest extends \PHPUnit\Framework\TestCase
     public function testExceptionHandlerAjaxDebugDisabledServiceUnavailable()
     {
         $request = new Request('/api/users/fetch');
-        $request->type('json');
+        $request->header('Accept', 'application/json');
         Router::request($request);
 
         $errorHandler = new MockErrorHandler();